WebAppServer/container-video-generator/requirements.txt
requests==2.31.0
boto3==1.34.0
mysql-connector-python==8.2.0
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
